pyproject.toml
[project]
name = "ContestTrade"
version = "0.1.0"
description = "ContestTrade: A Multi-Agent Trading System Based on Internal Contest Mechanism."
requires-python = ">=3.10"
dependencies = [
    "tqdm",
    "requests",
    "matplotlib",
    "tushare",
    "loguru",
    "openai",
    "tiktoken",
    "finnhub-python",
    "typer",
    "pydantic",
    "questionary",
    "PyYAML",
    "pandas",
    "langgraph",
    "rich",
    "tabulate",
    "yfinance>=0.2.65",
]
